MIX-281 fix: cancel orphaned server sessions when no local save exists and extract useGameIndex hook
